+++ src/App.jsx
@@ -12,6 +12,7 @@ import {
 	DictationExercise,
 	DraggableFillGaps,
 	InlineChoiceGroup,
+	LineMatch,
 	MemoryMatchGame,
 	RadioQuiz,
 	SelectExercise,
@@ -985,6 +986,14 @@ export default class App extends React.Component {
 						showDialog={this.showDialog}
 					/>
 				);
+			case "LineMatch":
+				return (
+					<LineMatch
+						config={value}
+						logError={this.logError}
+						showDialog={this.showDialog}
+					/>
+				);
 			case "Explanation":
 				return (
 					<>
@@ -1423,11 +1432,11 @@ export default class App extends React.Component {
 				);
 				break;
 			}
-			case "InlineChoiceGroup": {
-				articles.push(
-					<AccordionArticle
-						expandedByDefault={autoExpandSingleAccordion}
-						config={value}
+				case "InlineChoiceGroup": {
+					articles.push(
+						<AccordionArticle
+							expandedByDefault={autoExpandSingleAccordion}
+							config={value}
 						id={`${compoundID}-Accordion`}
 						key={`${compoundID}-Accordion`}
 						target={targetId}
@@ -1440,12 +1449,32 @@ export default class App extends React.Component {
 							showDialog={this.showDialog}
 						/>
 					</AccordionArticle>,
-				);
-				break;
-			}
-			case "Explanation": {
-				if (expandable) {
+					);
+					break;
+				}
+				case "LineMatch": {
 					articles.push(
+						<AccordionArticle
+							expandedByDefault={autoExpandSingleAccordion}
+							config={value}
+							id={`${compoundID}-Accordion`}
+							key={`${compoundID}-Accordion`}
+							target={targetId}
+							title={titleText}
+							titleHTML={titleTextHTML}
+						>
+							<LineMatch
+								config={value}
+								logError={this.logError}
+								showDialog={this.showDialog}
+							/>
+						</AccordionArticle>,
+					);
+					break;
+				}
+				case "Explanation": {
+					if (expandable) {
+						articles.push(
 						<AccordionArticle
 							expandedByDefault={autoExpandSingleAccordion}
 							config={value}